+++ src/decorators.ts
@@ -92,7 +92,7 @@ export function property(options?: PropertyOptions) {
 export function observe(...targets: string[]) {
   return (proto: any, propName: string): any => {
     if (!proto.constructor.hasOwnProperty('observers')) {
-      proto.constructor.observers = [];
+      Object.defineProperty(proto.constructor, 'observers', {value: []});
     }
     proto.constructor.observers.push(`${propName}(${targets.join(',')})`);
   }