+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased] 
 - Allow multiple decorators to be combined
+- Fix bug where `@observe` could not be used with `Polymer.mixinBehaviors`.
 
 ## [1.0.2] - 2018-01-26
 - Fix missing generated files from `1.0.1` release.